+++ tests/headers/unsigned.h
@@ -1,2 +1,3 @@
 #include <stddef.h>
 #include <stdint.h>
+wchar_t c;